broke structs out into functions
