+++ src/formats/internal/grid_version.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Clone)]
+pub enum GridVersion {
+    V6_2,
+    V6_1,
+    V6_0,
+}
