+++ src/formats/internal/rgb_color.rs
@@ -1,0 +1,12 @@
+#[derive(Debug, Clone)]
+pub struct RGBColor {
+    pub red: u8,
+    pub green: u8,
+    pub blue: u8,
+}
+
+impl RGBColor {
+    pub fn to_string(&self) -> String {
+        format!("#{:02x}{:02x}{:02x}", self.red, self.green, self.blue)
+    }
+}
