+++ src/formats/internal/rider.rs
@@ -1,0 +1,9 @@
+use crate::internal::vec2::Vec2;
+
+#[derive(Debug, Clone)]
+pub struct Rider {
+    pub start_position: Vec2,
+    pub start_velocity: Vec2,
+    pub start_angle: f64,
+    pub can_remount: bool,
+}
