+++ src/formats/internal/track_line/scenery.rs
@@ -1,0 +1,7 @@
+use crate::internal::track_line::Line;
+
+#[derive(Debug, Clone)]
+pub struct SceneryLine {
+    pub base_line: Line,
+    pub width: Option<f64>,
+}
