+++ src/formats/internal/track_line/simulation.rs
@@ -1,0 +1,10 @@
+use crate::internal::track_line::Line;
+
+#[derive(Debug, Clone)]
+pub struct SimulationLine {
+    pub base_line: Line,
+    pub flipped: bool,
+    pub left_extension: bool,
+    pub right_extension: bool,
+    pub multiplier: Option<f64>,
+}
