+++ src/formats/internal/trigger/background_color.rs
@@ -1,0 +1,18 @@
+use crate::internal::{rgb_color::RGBColor, trigger::FrameReachedEvent};
+
+#[derive(Debug, Clone)]
+pub struct State {
+    pub color: RGBColor,
+}
+
+#[derive(Debug, Clone)]
+pub struct Trigger {
+    pub state: State,
+    pub trigger_event: FrameReachedEvent,
+}
+
+#[derive(Debug, Clone)]
+pub struct TriggerGroup {
+    pub initial_state: State,
+    pub triggers: Vec<Trigger>,
+}
