+++ src/formats/internal/vec2.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone)]
+pub struct Vec2 {
+    pub x: f64,
+    pub y: f64,
+}
